backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.8.1
boto3==1.35.63
botocore==1.35.63
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.3
django-cors-headers==4.7.0
django-filter==25.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.30.0
google-generativeai==0.8.5
googleapis-common-protos==1.68.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
inflection==0.5.1
jmespath==1.0.1
numpy==2.3.2
packaging==24.2
pillow==11.0.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.0
pyparsing==3.2.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
sentence-transformers==3.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
websockets==15.0.1
whitenoise==6.9.0
